Add ability to save video from a generator
Since large video will not fit in memory, this ability allows us to read one
frame from the video file, transform the frame and write it to another video
file. Thus there is only need for one frame at a time in memory.

Also adds a pipeline node to show how it is done.